backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
azure-cosmos==4.5.1
openai>=1.50.0
celery[redis]==5.3.6
redis[hiredis]==5.0.1
apscheduler==3.10.4
moviepy==1.0.3
Pillow==10.1.0
requests==2.31.0
python-multipart==0.0.6
aiofiles==23.2.1
urllib3==2.1.0
pyjwt==2.8.0
bcrypt==4.1.2
azure-storage-blob==12.19.0
google-api-python-client==2.111.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
